refactor(header): Header 컴포넌트 구조 리팩터링 및 하위 컴포넌트 분리
- BrandHeader : 로고를 통해 홈으로 이동할 수 있게 하는 링크 컴포넌트
- LanguageButton : 현재 언어를 보여주고, 언어 변경 인터페이스로 진입하는 버튼 컴포넌트, 디자인만 구현
- LoginButton : 로그인 동작을 트리거하는 버튼 컴포넌트
- NavBarLeft : 강의, 로드맵, 멘토링, 커뮤니티 등 주요 서비스로 이동할 수 있는 상단 내비게이션 메뉴 컴포넌트
- NavBarRight : 대시보드, 재생 버튼, 수강바구니, 프로필 등 로그인 후 우측 상단에 표시되는 사용자 인터페이스 컴포넌트
- SearchInout : 실무 강의를 검색할 수 있도록 입력창과 검색 아이콘을 제공하는 검색 입력 컴포넌트
