+++ src/layout/Header/LoginButton.jsx
@@ -1,0 +1,15 @@
+import { Button } from '@mantine/core';
+
+function LoginButton({ onLogin }) {
+
+    return <Button
+        onClick={onLogin}
+        variant="filled"
+        size="xs"
+        color="green"
+    >
+        로그인
+    </Button>
+}
+
+export default LoginButton;
